Plot method for MAMS.stepdown objects

produces as plot of the boundaries.

Usage

# S3 method for class 'MAMS.stepdown'
plot(
  x,
  col = NULL,
  pch = NULL,
  lty = NULL,
  main = NULL,
  xlab = "Analysis",
  ylab = "Test statistic",
  ylim = NULL,
  type = NULL,
  bty = "n",
  las = 1,
  ...
)

Arguments

x

An output object of class MAMS.stepdown

col

A specification for the default plotting color (default=NULL). See par for more details.

pch

Either an integer specifying a symbol or a single character to be used as the default in plotting points (default=NULL). See par for more details.

lty

A specification for the default line type to be used between analyses (default=NULL). Setting to zero suppresses plotting of the lines. See par for more details.

main

An overall title for the plot (default=NULL).

xlab

A title for the x axis (default="Analysis").

ylab

A title for the y axis (default=`"Test statistic"“).

ylim

A title for the y axis (default="Test statistic").

type

Type of plot to be used (default=NULL). See plot for more details.

bty

Should a box be drawn around the legend? The default "n" does not draw a box, the alternative option "o" does.

las

A specification of the axis labeling style. The default 1 ensures the labels are always horizontal. See ?par for details.

...

Further arguments passed to or from other methods.

Value

Graphic output.

Author

Thomas Jaki, Dominique-Laurent Couturier

Examples

if (FALSE) { # \dontrun{
# 2-stage design with triangular boundaries
res <- mams(K=4, J=2, alpha=0.05, power=0.9, r=1:2, r0=1:2,
 p=0.65, p0=0.55, ushape="triangular", lshape="triangular", nstart=30)

plot(res)
} # }