Function to find stopping boundaries for a 2- or 3-stage (step-down) multiple-comparisons-with-control test.

The function determines stopping boundaries for all intersection hypothesis tests in a multi-arm multi-stage study, given the amount of alpha (familywise error rate) to be spent at each analysis.

Usage

stepdown.mams(nMat=matrix(c(10, 20), nrow=2, ncol=4),
             alpha.star=c(0.01, 0.025), lb=0,
             selection="all.promising")

Arguments

nMat

Matrix containing the cumulative sample sizes in each treatment arm columns: control, trt 1, ..., trt K), at each analysis (rows). The number of analyses must be either 2 or 3 (default=matrix(c(10, 20), nrow=2, ncol=4)).

alpha.star

Cumulative familywise error rate to be spent at each analysis (default=c(0.01, 0.025)).

lb

Fixed lower boundary (default=0).

selection

How are treatments selected for the next stage? Using the default "all.promising" method, all treatments with a test statistic exceeding the lower boundary are taken forward to the next stage. If "select.best", only the treatment with the largest statistic may be selected for future stages. (default="all.promising").

Value

An object of the class MAMS.stepdown containing the following components:

l

Lower boundaries.

u

Upper boundaries.

nMat

Cumulative sample sizes on each treatment arm.

K

Number of experimental treatments.

J

Number of stages in the trial.

alpha.star

Cumulative familywise error rate spent at each analysis.

selection

Pre-specified method of treatment selection.

zscores

A list containing the observed test statistics at analyses so far (at the design stage this is NULL).

selected.trts

A list containing the treatments selected for each stage.

Details

The function implements the methods described in Magirr et al (2014) to find individual boundaries for all intersection hypotheses.

References

Jaki T., Pallmann P. and Magirr D. (2019), The R Package MAMS for Designing Multi-Arm Multi-Stage Clinical Trials, Journal of Statistical Software, 88(4), 1-25. Link: doi:10.18637/jss.v088.i04

Magirr D., Jaki T. and Whitehead J. (2012), A generalized Dunnett test for multi-arm multi-stage clinical studies with treatment selection, Biometrika, 99(2), 494-501. Link: doi:10.1093/biomet/ass002

Magirr D., Stallard N. and Jaki T. (2014), Flexible sequential designs for multi-arm clinical trials, Statistics in Medicine, 33(19), 3269-3279. Link: doi:10.1002/sim.6183

Stallard N. and Todd S. (2003), Sequential designs for phase III clinical trials incorporating treatment selection, Statistics in Medicine, 22(5), 689-703.

Author

Dominic Magirr

Examples

# \donttest{
# Note that some of these examples may take a few minutes to run
# 2-stage 3-treatments versus control design, all promising treatments
# are selected:
stepdown.mams(nMat=matrix(c(10, 20), nrow=2, ncol=4), 
              alpha.star=c(0.01, 0.05), lb=0, 
              selection="all.promising")
#> Design parameters for a 2 stage trial with 3 treatments
#> 
#>                                                   Stage 1 Stage 2
#> Cumulative sample size  (control):                     10      20
#> Cumulative sample size per stage (treatment  1 ):      10      20
#> Cumulative sample size per stage (treatment  2 ):      10      20
#> Cumulative sample size per stage (treatment  3 ):      10      20
#> 
#> Maximum total sample size:  80 
#> 
#> 
#> Intersection hypothesis H_{ 1 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.33    1.67
#> Lower boundary       0.00    1.67
#> 
#> Intersection hypothesis H_{ 2 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.33    1.67
#> Lower boundary       0.00    1.67
#> 
#> Intersection hypothesis H_{ 1 2 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.56    1.95
#> Lower boundary       0.00    1.95
#> 
#> Intersection hypothesis H_{ 3 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.33    1.67
#> Lower boundary       0.00    1.67
#> 
#> Intersection hypothesis H_{ 1 3 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.56    1.95
#> Lower boundary       0.00    1.95
#> 
#> Intersection hypothesis H_{ 2 3 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.56    1.95
#> Lower boundary       0.00    1.95
#> 
#> Intersection hypothesis H_{ 1 2 3 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.68    2.10
#> Lower boundary       0.00    2.10
# select the best treatment after the first stage:
stepdown.mams(nMat=matrix(c(10, 20), nrow=2, ncol=4), 
              alpha.star=c(0.01, 0.05), lb=0, 
              selection="select.best")
#> Design parameters for a 2 stage trial with 3 treatments
#> 
#>                                                   Stage 1 Stage 2
#> Cumulative sample size  (control):                     10      20
#> Cumulative sample size per stage (treatment  1 ):      10      20
#> Cumulative sample size per stage (treatment  2 ):      10      20
#> Cumulative sample size per stage (treatment  3 ):      10      20
#> 
#> Maximum total sample size:  80 
#> 
#> 
#> Intersection hypothesis H_{ 1 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.33    1.67
#> Lower boundary       0.00    1.67
#> 
#> Intersection hypothesis H_{ 2 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.33    1.67
#> Lower boundary       0.00    1.67
#> 
#> Intersection hypothesis H_{ 1 2 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.56    1.90
#> Lower boundary       0.00    1.90
#> 
#> Intersection hypothesis H_{ 3 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.33    1.67
#> Lower boundary       0.00    1.67
#> 
#> Intersection hypothesis H_{ 1 3 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.56    1.90
#> Lower boundary       0.00    1.90
#> 
#> Intersection hypothesis H_{ 2 3 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.56    1.90
#> Lower boundary       0.00    1.90
#> 
#> Intersection hypothesis H_{ 1 2 3 }: 
#> 
#>                   Stage 1 Stage 2
#> Conditional error    0.01    0.05
#> Upper boundary       2.68    2.02
#> Lower boundary       0.00    2.02
# 3 stages and unequal randomization:
stepdown.mams(nMat=matrix(c(20, 40, 60, rep(c(10, 20, 30), 3)), 
              nrow=3, ncol=4), 
              alpha.star=c(0.01, 0.025, 0.05), lb=c(0, 0.75), 
              selection="all.promising")
#> Design parameters for a 3 stage trial with 3 treatments
#> 
#>                                                   Stage 1 Stage 2 Stage 3
#> Cumulative sample size  (control):                     20      40      60
#> Cumulative sample size per stage (treatment  1 ):      10      20      30
#> Cumulative sample size per stage (treatment  2 ):      10      20      30
#> Cumulative sample size per stage (treatment  3 ):      10      20      30
#> 
#> Maximum total sample size:  150 
#> 
#> 
#> Intersection hypothesis H_{ 1 }: 
#> 
#>                   Stage 1 Stage 2 Stage 3
#> Conditional error    0.01   0.025    0.05
#> Upper boundary       2.33   2.070    1.70
#> Lower boundary       0.00   0.750    1.70
#> 
#> Intersection hypothesis H_{ 2 }: 
#> 
#>                   Stage 1 Stage 2 Stage 3
#> Conditional error    0.01   0.025    0.05
#> Upper boundary       2.33   2.070    1.70
#> Lower boundary       0.00   0.750    1.70
#> 
#> Intersection hypothesis H_{ 1 2 }: 
#> 
#>                   Stage 1 Stage 2 Stage 3
#> Conditional error    0.01   0.025    0.05
#> Upper boundary       2.57   2.340    2.03
#> Lower boundary       0.00   0.750    2.03
#> 
#> Intersection hypothesis H_{ 3 }: 
#> 
#>                   Stage 1 Stage 2 Stage 3
#> Conditional error    0.01   0.025    0.05
#> Upper boundary       2.33   2.070    1.70
#> Lower boundary       0.00   0.750    1.70
#> 
#> Intersection hypothesis H_{ 1 3 }: 
#> 
#>                   Stage 1 Stage 2 Stage 3
#> Conditional error    0.01   0.025    0.05
#> Upper boundary       2.57   2.340    2.03
#> Lower boundary       0.00   0.750    2.03
#> 
#> Intersection hypothesis H_{ 2 3 }: 
#> 
#>                   Stage 1 Stage 2 Stage 3
#> Conditional error    0.01   0.025    0.05
#> Upper boundary       2.57   2.340    2.03
#> Lower boundary       0.00   0.750    2.03
#> 
#> Intersection hypothesis H_{ 1 2 3 }: 
#> 
#>                   Stage 1 Stage 2 Stage 3
#> Conditional error    0.01   0.025    0.05
#> Upper boundary       2.70   2.490    2.20
#> Lower boundary       0.00   0.750    2.20
# }